+++ src/store/features/templates/reducer.ts
@@ -10,10 +10,7 @@ export const templatesReducer = createReducer(initialState, (builder) => {
     // template edits do not return the current columns, so they are re-added to avoid losing them
     .addCase(editTemplate.fulfilled, (state, action) =>
       state.map(
-        (t) =>
-          t.id === action.payload.id
-            ? {...action.payload, columns: t.columns} // keep columns, change the rest
-            : t // only change edited templates, other stay the same
+        (t) => (t.id === action.payload.id ? {...action.payload} : t) // only change edited templates, other stay the same
       )
     );
 });
